+++ client/package.json
@@ -34,6 +34,7 @@
     "react-slick": "^0.29.0",
     "recoil": "^0.7.6",
     "slick-carousel": "^1.8.1",
+    "socket.io-client": "^4.5.4",
     "styled-components": "^5.3.6",
     "typescript": "^4.8.4",
     "web-vitals": "^2.1.4"
