+++ client/src/pages/HomePage/HomePage.tsx
@@ -1,8 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import React, { Ref, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { InfiniteData, useInfiniteQuery, useQuery } from 'react-query';
-import { AxiosError } from 'axios';
 import { useInView } from 'react-intersection-observer';
 import { useSetRecoilState } from 'recoil';
 import user from '../../store/userAtom';
@@ -16,9 +14,9 @@ import BoardItem from '../../components/BoardItem/BoardItem';
 // style
 import S from './HomePageStyles';
 // type
-import { Api, Board } from '../../types/responseData';
-// api
-import { getBoardData } from '../../apis/api/boardApi';
+import { Board } from '../../types/responseData';
+// hook
+import useBoardQuery from '../../hooks/useBoardQuery';
 
 const makeBoarList = (boards: Board[]) => {
   const boardList = boards.map((board) => {
@@ -29,9 +27,7 @@ const makeBoarList = (boards: Board[]) => {
 
 const HomePage = () => {
   const [ref, isView] = useInView();
-  const setUserData = useSetRecoilState(user);
   const navigate = useNavigate();
-  const requestCount = 10000;
   const addPostButton = {
     buttonImg: addPostButtonImg,
     eventHandler: () => {
@@ -40,59 +36,7 @@ const HomePage = () => {
     description: '게시물을 추가할래요.',
   };
 
-  // const { data: boards } = useQuery<Board[], AxiosError<Api>>(
-  //   'boards',
-  //   () =>
-  //     getBoardData(requestCount, '-1').then((response) => response.data.boards),
-  //   {
-  //     onError: (error) => {
-  //       if (
-  //         error.response &&
-  //         error.response.data.statusCode === 401 &&
-  //         error.response.data.error === 'Unauthorized'
-  //       ) {
-  //         setUserData({ userId: -1, userName: '', userProfileUrl: '' });
-  //         navigate('/');
-  //       }
-  //     },
-  //   },
-  // );
-
-  // 테스트 중입니다.
-  const fetchBoardList = async ({ pageParam = -1 }) => {
-    const response = await getBoardData(4, String(pageParam));
-    const result = response.data;
-    const nextPage = result.next_max_id;
-    return {
-      result: result.boards,
-      nextPage,
-      isLast: nextPage < 1,
-    };
-  };
-
-  const query = useInfiniteQuery<
-    { result: [Board]; nextPage: number; isLast: boolean },
-    AxiosError<Api>
-  >('[boardlist]', fetchBoardList, {
-    getNextPageParam: (lastPage, pages) => {
-      if (!lastPage.isLast) return lastPage.nextPage;
-      return undefined;
-    },
-    refetchOnWindowFocus: false,
-    refetchOnMount: true,
-    refetchOnReconnect: true,
-    retry: 1,
-    onError: (error) => {
-      if (
-        error.response &&
-        error.response.data.statusCode === 401 &&
-        error.response.data.error === 'Unauthorized'
-      ) {
-        setUserData({ userId: -1, userName: '', userProfileUrl: '' });
-        navigate('/');
-      }
-    },
-  });
+  const query = useBoardQuery();
 
   useEffect(() => {
     if (isView && query.hasNextPage) {
